config/requirements.txt
# StockPulse Backend — Python Dependencies
supabase>=2.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
numpy>=1.26.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
